Add precise priming synchronization
For CSQC enabled clients we synchronize the grenade timer with the
server's notion of when the grenade was actually primed to eliminate
ping dependency.

Augment this by transmitting the actual time the grenade was primed.
This both improves accuracy and allows future client-side physics for
grenade jump emulation.

We introduce a temporary fo_grentimer == 3 mode which uses the new
prediction for the offset to prime, while maintaining RTT/2 for
throwgren.