Implement precise grenade throwing
Up to this point, all version of Team Fortress (including FortressOne)
have implemented grenade throwing using a test which executes every
100ms.  This introduces a random input-delay (of up to 100ms) dependent
on when you press throw relative to the next timer evaluation.

This introduces a new, 'precise' timing mode that throws grenades
immediately when instructed.

We reserve the final 100ms before the grenade explodes.  Previously, it
was random whether or not you would be able to throw in this time.  Now
it is deterministic, if you hold a grenade to this point, you are along
for the ride.  This is to allow future client-side prediction on
knockback to function reliably (as this allows us to predict the grenade
explosion position from player position when held more reliably).

Note that this also explains other behaviors, such as inconsistency in
airblast hitting thrown grenades when tightly timed.  In this case, the
input lag on throw was allowing the grenade throw to actually be
re-ordered with the airblast, so that it was thrown after -- even when
pressed before.

To enable:
  setinfo precise_grenades on
or
  setinfo pg on