Fix not-exploding napalm grenades (and eliminate grentimer effects)
A long standing bug is that napalm grenades occasionally do not explode
properly, leaving glowing, stuck entities.

This turns out to be a race between several pieces of the code.

1. When the grenade initially explodes render a flare and start a timer
   program for the subsequent napalm explosions ("grentimer").
2. When the program in (1) begins 100ms later, it tries to track the
   number of napalm (or gas) grenades active for the player in question.
3. When the player dies, conditional on there being no active napalm
   grenades we try to reclaim these timers.

It turns out that when (3) occurs between the 100ms between (1) and (2)
that the accounting can be incorrect and we'll remove the timer, but
this incorrectly eliminated the subsequent explosion and left the
grenade pinned until the ent was reclaimed (since the timer also owned
its removal).

There is no need for either:
  (1) The extra timer entity
  (2) The existing spam tracking

For (1) we can extend the think state machine on the existing entity for
both napalm and gas, this also allows us to completely eliminate all the
clean-up complexity that (3) adds [now we can just remove grenades,
rather than worrying about grenades also parenting area of effect timers
that we also have to remove].

We throw (2) out for now.  There are cleaner ways to write it if it's
needed and it wasn't even working for gas grenades (because someone
changed the timer name at some point).