Grenade ping correction
Due to their <prime> <wait> <throw> nature, grenades have all sorts of
inconsistencies introduced by delay between client and server around
each event.

Two players that see each other at the same time, and prime at the same
time, will actually not be able to throw at the same time as the
player with the lower ping is able to get the server to register their
prime first.

This is particularly material since grenades don't allow you to throw
for the first 800ms, so even 50-100ms ping differences are a large
relative advantage.

This introduces ping correction, so that the point at which we consider
the client to have primed is corrected for their latency when enabled.
We ensure that this does not let players prime any faster, or more
frequently, than they would have been able to before, only more
consistently.

One very nice property is that unlike projectile projection, which
causes visible projectiles to skip, there are no strange artifacts for
other clients as the adjustment occurs before it's possible to throw the
grenade.

Also rewrite the grenade timer rendering client code to be more compact
and to handle the fact that the expiry offset has been adjusted to
correct for latency.

Clients that do not support CSQC can simply playsound at the point which
they prime; this will now be correctly aligned (as the point they primed
is properly reflected to the server).

Controlled by:
  localinfo precise_grenades on/off

While we'd normally default a server setting on, this is a big enough
change that we'll give it a pug or two of testing before pushing it as a
default to soak.  [For this reason we also temporarily allow mid-round
changing of the tunable.]