The hunt continues.
So a major part of the issue seems to be that the reference "self" was
changing due to migration of timers into a class.  However, it turns out
we were getting fortunate before and that self was actually the world
entity and not the player.  This turns out to be the entire reason that
soundupdate even worked with localsound, as the latter implies a default
entity of zero.  This explains why we got double sounds even with the
identical calls under the new code (as implicit entities from the
grentimer entities were substituted.

This seems to work, I get very occasional misses, but I suspect those
were the ones you were also seeing before.

I'm not sure about the PHS stuff, I've gone through the FTE code and it
doesn't look like it's strongly interpreted for Q1 despite the docs and
other mod's usage.  I've restored the extra channels for now.

I've left the old sound code as the default while I test the new one in
a few pugs before flipping them.

Also made sure the timer sound was cached to avoid potential hitching
(and anything to generally increase reliability).