Connect detpipe to rewind
We use the rewound position to calculate damage, but the current
position to calculate knockback.

This will bear some tinkering as movement replay gets more sophisticated
(as well as rewinding some flag events), but this is OK as a testing
starting point.

We have to be careful of a few things, e.g. not applying to rewind to
non-pipebombs as they share think programs as well as not rewinding
server triggered events (such as death detonating pipebombs) since
they're independent of user latnecy.