Add client_time
Rather than implementing a-la-carte ping correction, introduce a centrally
updated notion of client_time, that we can use in place of time when calculating
next think times.  We can also simplify all of the places we are careful not to
go backwards in time by ensuring it's monotonically increasing.

Update weapon projection to be based on client_time.  Also document some errors
in the current math that we'll forrect in the future.