Move weapon projection to use client_time
The original code followed the reasoning:

`Project projectiles forward in time based on ping up to configurable
 maximum (default 100ms).
 Effective projection time is maximum / 2 (eg 50ms), because we assume RTT
 is evenly distributed in both directions - and we want the server to spawn the
 entity in the position it would have have been if it had received the message
 (player_ping / 2) ago.`

This was is not quite correct.  It's true the client fired rtt/2 ago, however,
this is not the whole story.   The point that client fired from an was also an
interpolated point at time-rtt/2.