Add weapon prediction infrastructure
Create a per-player entity for storing and communicating weapon-prediction
state.  Introduce a hybrid-driven client_think() that can be used to synchronize
and initiate animation.

Much of this infrastructure is inspired by the ktx/reki implementations.

Annoying subtlety: We use (deprecated) visible dimension stuff for the flash
code, if this isn't set, no one gets updates on the entity.