Migrate to autocvar and update default fo_grentimer
Both because cvar string interaction is suspect in a crash and because
autocvars are just better, migrate all* of our cvars to use autocvar.

Also update fo_grentimer default to 2, it had not been updated to
reflect the work on ping correction (either new or old).

*: Some of the menu code uses dynamically constructed names which are
incompatible with autocvar so we exclude those.