Ensure ping correction can't affect firing rate
In adjusting attack_finished for ping we need to make this on both sides, e.g.
when the client fires and when we subsequently check.  In missing the second
we'd allow firing rate to increase with latency.  Also re-add catch-all that was
previosuly present for no ammo and minor cleanup.