Tie attack_finished to client_time() for now
The prior approach failed to account for the fact that some of weapon animations
have attack_finished embedded in them, resulting in overlapping calls.  This was
pushing the finished time out unreasonably.

Just remove the max for now, we're going to make some other changes that will
smooth out client_time in the face of jitter anyway.