Harden interfaces with opaque Slot tokens
We've seen one or two crashes with out of bound slots, the code is also
confusing in terms of trying to determine when you need to do input checking and
when you can just trust that a slot is valid.  Try to better formalize this by
introducing an opaque slot token type that references valid slots.