Introduce stable clock
`time` turns out to not be particularly well-aligned with frames, with it being
often +/- half a frame by timing.  Improve this by introducing a stable
client_time clock (which subsumes some of the prior notion).  Infrastructure for
aligning think events (such as animation) with this clock (which can then be
more strongly aligned with the client) is added.

While we're here, migrate a bunch of other things that we're going to need to
synchronize with this clock (too much really, but hard to split now).  At least
we simplified the reload timing and tracking in the transition.

We pull a bunch of things back to time (e.g. pipecooldown) are not migrated yet
since we need to also unify client_time with rewind or just generally test and
tune.