Fix reload impulse blocking
Pulling reload out of a timer was leading to an early return that blocked some
impulses.  This was not obvious since the new button commands directly invoke
them and were not blocked.  Fix this.

We should also make the buttons trigger the standard impulse paths so this
discrepancy isn't possible in the future (particularly the inverse, a command
that should be blocked by reload, but isn't.)