Implement fixed-time physics in QC
Re-implements FLY and BOUNCE physics in quakec. This is both necessary for us to
be able to implement predicted grenade projectiles (which follow arcs) but also
allows us to better align our client and server side predictions.  We do this by
moving physics to a fixed step (excess in the step is discarded) that can be
aligned to ensure constant progression for a given time between client and
server.  The net-net is that this avoids small computational differences
resulting in different bounces for client vs server side (since difference in
time -> difference in acceleration -> difference in bounce/cutoff).

One immediate benefit: antilag forwarding of grenades is now consistent with
true grenade arc, as opposed to the linear projection prior.

This is always used for client-side prediction and antilag projection.  When
"qc_physics 1" (or qcp) is set we also use it server side (as a customphysics)
handler.