Turn on weapon prediction by default
Originally was separating this but it's behaving fairly well with latest updates
and turns out to actually matter for projectiles (since otherwise you can change
weapons fast enough that client still thinks you are using previous weapon and
creates incorrect predicted projectiles).

Could possibly separate tracking and rendering in the future but goal is to just
make this unnecessary and always automatically on.