Introduce HUD render cache
The HUD code allows a fairly sophisticated extent of customization, however it
is expensive to compute the end product of all of these knobs on a per-frame
state.  Instead of rendering from scratch every time, introduce a render time
and an FPS target.  For frames out of the FPS target we use the cached
rendering.

Improves average HUD render time to 1ms