Implement partial rendering
Now that we have a render cache, we can apply a very good trick: we don't need
to render all of the hud every frame.  Instead, implement a per-element timing
and FPS (controlled by fo_hud_fps) so that we only render one element per frame
-- the oldest element that needs rendering.  If no elements need rendering, then
do nothing.

This improves average HUD render time to 0.07ms, with a maximum observed time of
0.2ms.

FPS with moderate geometry improves by ~400-500fps on my machine, with simple
geometry >1000fps.  Even though people play with an FPS cap, the important
change is that the frame latency reduction reduces the sensation of mouse
hitches when there are expensive hud elements.