Move newmis behaviors to quakec
Using 'newmis' to initialize an entity triggers special handling in qwsv that
forwards it for 50ms arbitrarily.  This is awkward to rectify with projectile
physics as it does it with a 50ms physics window.  Move this behavior into
quakec and split the 50ms into a fixed and dynamic period, so that we can open
up the range at which predictiles feel uniform slightly.