Make newmis (static) nudge a tunable
The old physics code would additionally apply an (inconsistent) outstanding
physics frames to new projectiles.  The new code is more consistent but add a
tunable for extra nudge in case we want to restore some of the boost that this
was previously providing.  It's small enough that this might not matter.