Reduce bandwidth used by ~25%
Velocity was embedded into a stat field for speed display on the HUD.  This is a
non-optimal choice both because it changes incredibly quickly and a better (more
up-to-date) copy is already available client side in the form of pmove.

Reduces BW used from ~60k/s to 45k/s.  Should also likely help with proxy
overflows.