Augment, don't override flags when popping up shock grenades
We're seeing the occasional disconnect where a projectile is missing its type.
But this should always be sent; seems to happen to HPBs.  One guess is maybe the
original packet is being lost on a shock grenade and then flags is being smashed
by the movetype update?  Ensure this doesn't happen by ORing (and make sure we
actually pass back the movetype -- didn't matter since no vel).

In the meantime, prevent it from crashing by giving it a (intentionally
probably wrong) model.