Fix autocvars around prior FOCMD defines
This was generating compilable, but not correct code that would end up using
FOCMD_... for cvar definitions.  This was further masked by an incorrect use of
cvar() sneaking back into the code that allowed it to pick up the legacy value
on a client that had previously defined it (although it would be missing on a
new install).

While we're here, shoot another cvar() that's crept in around fov.