Add wpp_setspeed which allows CSQC to implement speed updates rather than stuff
This reduces lerping/prediction errors for locally implemented weapons as we do
not need wait for the server to initiate TFSTATE_AIMING, etc.  This also avoids
the use of lower maxspeed values, which do not arrive in time for pmove --
resulting in large lerps on state transition.

Also:
- Remove legacy independent tracking of forward/back and sidespeed, which was
always symmetric and mirrored into qwsv's maxspeed.

- Some cleanup of wpp_status() that snuck in while refactoring.