weapons: fix mashing +slot
mashing +slot on top of +slot creates confusion since we couldn't actually
process the prior weapon change from -slot even though a new +slot is queuing
and overloading it.  This resulted in the state-machine valid, but confusing
behavior that if you +slot before a prior -slot attack finished, then it might
consider the weapon to swap back to on the second +slot as the original.  Make
this more intuitive by making sure we forward the prior -slot to last-weapon
before this.

probably worth improving impulse behavior immediately after this by using
queue_slot to eliminate a bunch of cases for these held impulses that can be
clobbered.