prediction: use localsound and tie to projectile
Use localsound to avoid attenuation that was occurring with predicted location.
Now that we have projectile prediction for everything tie tracking of predicted
sounds to the projectile instead of a heuristic.

We're not doing this for hitscan weapons, but technically we werent before
anyway since they had no projectile to trigger the missed sound check.