experimental: restore air control / strafing
Air control with the new conc was harder due to the way it was perturbing
forward (and strafe moves perpendicular from this).  A nice compromise is that
we only need to perturb the vector while firing; this might look a little weird
while spectating (we can improve this later) but we can only apply the effect
while firing, which allows regular air control while conc jumping (and not
firing).

Setting fo_concuss 1 will now enable the following defaults:
  air control, affects medic

Also add more aggressive resets to make sure we clear crosshair when not conced,
add `fo_crossy` for people who want to set cl_crossy now that we're overriding
it.