conc: massive cleanup
Now that we have so many implementations floating around, all of the state
around how we heal, do timers, clear on cap, etc is a giant mess.  Even figuring
out what the conc type is requires a multi-way eval.

Refactor all of this so it goes through a few common functions.

All so we can have bubbles on the new conc.  The long form name enable name is
also now fixed to be `fo_concuss`.