login: fix file-handle leaks and refactor
The current login code is leaking file handles in two places which leads
to login failures when we run out of handles and it's unable to read the
token.

Fix this but also just refactor so we're not duplicating the read/send
path between client and server initiated.  Add instructions for people
who dont have a token to visit the FO website to get one.