init: no more stuff
If you've ever noticed that first connect seems to take a long time compared to
subsequent ones it turns out that hundreds of mod aliases (many of which we no
longer really use) are being stuffed (with a rate limit) for binding.

Eliminate this by performing client binds directly in CSQC, with no server-side
stuffs.