spec: render grenade timer and play sounds
Deliver grenade information to spectators* so that the hud grenade timers are
synchronized with the watched player.

*: technically removing _NOSPEC should have gone a lot of the way with the prior
implementation, however these don't seem to work correctly for unknown reasons.
We want this for TFX stuff anyway so just try the broadcast approach.  We should
be able to avoid need for reliable transmit with local gren timers also.