weapons: make weapon changing more reliable and add new features
When queued on attack_finished we leave other impulses like weapon changes on
the line until it is finished, making it possible for them to be overwritten
when more than one impulse is sent.

Improve this by moving weapon impulses (except quickslot because it's difficult
to do the right thing there) to immediately queue rather than stay pending so
that we can process them more reliably.

Also add two new client options:
  `fo_nofirealpha` (default: -1) when set (e.g. not -1) controls alpha when you
    can't fire
  `fo_hue_weaponswap` (default: 0) when set, allows you to visually swap weapons
    prior to attack finished.
Note that these can be combined.