antilag: rewind for death projectiles
It's possible to die in between the time that you fire and the server receives
you firing.  This is particularly frustrating since you can often see the client
side projectile disappear as a result of the server processing your death first.

Handle this by introducing limited reconciliation that can emit projectiles in
this window.  Needs more integration with rewind and hitscan but good enough for
a first try.