wp: fix state tear
When we refactored how the prediction update calculation occurred it could miss
elements in think, allowing a handled impulse to propagate to the next frame.
This ended up being typically invisible because it wouldn't be an effect-frame,
but with very low pings it was possible for this to occur; resulting in effects
like some double prints and double projectiles.