pm: prediction for held grenades
This is actually interesting since the entity associated with the knock only
exists for 100ms.  Meaning for pings over 100ms the client never sees the
grenade before it's exploded.  Handle this by treating the held grenade like a
locally fired projectile and piggy-backing on that infrastructure, allowing us
to forward simulate the virtual projectile and emit a knock we'll later stitch
up with the actual ent.