new-balance: new take on engineer's rail gun
A new attempt at railgun for the engineer, aims to add mobility to the class and
add a better mechanism for catching escaping players.

It still fires a rail, but it fires backwards -- from the aimed at wall, back to
the engineer.  This can be used to either intercept someone escape or for the
engineer to gain additional mobility (jump higher, move faster).